ajustes e inclusao de VendaService, VendaEntity, VendaRepository, Venda Controller e VendaDTO

+++ src/main/java/br/edu/atitus/api_example/dtos/ImovelDTO.java
@@ -1,7 +1,9 @@
 package br.edu.atitus.api_example.dtos;
 
+import java.math.BigDecimal;
+
 public record ImovelDTO(String nome, String logradouro, String numero, String bairro, String complemento, String cep,
 		String cidade, String estado, double area, double latitude, double longitude, String descricao, String tipo,
-		double valor) {
+		BigDecimal valor) {
 
 }
